+++ api/src/main/java/net/jqwik/api/providers/TypeUsage.java
@@ -32,6 +32,7 @@ import static org.apiguardian.api.API.Status.*;
  * </ul>
  */
 @API(status = MAINTAINED, since = "1.0")
+@NonNullApi
 public interface TypeUsage {
 
 	@API(status = INTERNAL)