+++ kotlin/src/test/kotlin/net/jqwik/kotlin/api/NullabilityTests.kt
@@ -1,9 +1,6 @@
 package net.jqwik.kotlin.api
 
-import net.jqwik.api.Arbitraries
-import net.jqwik.api.Arbitrary
-import net.jqwik.api.Example
-import net.jqwik.api.ForAll
+import net.jqwik.api.*
 import net.jqwik.testing.TestingSupport
 import java.util.*
 
@@ -22,5 +19,13 @@ class NullabilityTests {
             nullableArbitrary.generator(1000),
             random
         ) { s: String? -> s == null }
+    }
+
+    @Example
+    fun test(@ForAll aString : String, @ForAll aNullableString: String?) {
+        val parameters = NullabilityTests::test.parameters
+        for (parameter in parameters) {
+            println(parameter.type.isMarkedNullable)
+        }
     }
 }